clean up traffic weights
